Add Creality Ender-3 BTT SKR E3 DIP v1.x config
This adds a set of configuration files for stock
Creality Ender 3 printer retrofitted with a BigTreeTech's
SKR E3 DIP v1.0 or v1.1 board. No changes need to be made
for the v1.1 board as it only differs in SERVOS header pinout.

This configuration assumes that A4988 stepsticks are used
for all the axes including the extruder.

Signed-off-by: Alexander Amelkin <alexander@amelkin.msk.ru>