Correct breakage of FFCP config post PR 22983
PR 22983 on the Marlin bugfix-2.0.x repository removed the defines previously assigned to the extruder fan pins. Although this brought the MightyBoard more in line with the behavior of other boards, it breaks the existing configuration file for the FFCP and the generic assignments aren't compatible with the factory-shipped board in this machine.

This commit fixes the break by directly assigning the pins used on the FFCP powered by the Rev. E (and Rev. D) Mightyboard. As such, the configuration files will be compatible with the Marlin repo both prior to and after PR 22983.