Add CR-10S Pro
This commits adds the default configuration for the CR-10S Pro with an
stock capcitiv or resistiv probe (not BLtouch!).
It is a good starting point for further tweaks for the specific
printer e. g.:
- optize machine limits
- tweak bed size
- tweak leveling settings

This is a working configuration, above mentioned tweaks are only a
suggestion for further tinkering.